(0716) feat: #8 특정 회원 조회 기초 로직 완성 (1616)
