+++ MAPICK/src/main/java/com/gitsunjaeab/mapick/api/member/dto/request/MemberUpdateRequest.java
@@ -1,16 +1,20 @@
-package com.gitsunjaeab.mapick.api.member.dto;
+package com.gitsunjaeab.mapick.api.member.dto.request;
 
+import com.gitsunjaeab.mapick.api.member.dto.MemberEmailUnique;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 
 @Getter
 @Setter
-public class MemberRequest {
+@NoArgsConstructor
+public class MemberUpdateRequest {
 
-//    private Long id;
+    @NotNull
+    private boolean isBlacklisted;
 
     @Size(max = 255)
     private String name;
@@ -33,6 +37,13 @@ public class MemberRequest {
     @Size(max = 255)
     private String provider;
 
+    @NotNull
+    @Size(max = 255)
+    private String role;
+
+    @Size(max = 255)
+    private String status;
+
     @Size(max = 255)
     private String profileImage;
-}
+}
