+++ MAPICK/src/main/java/com/gitsunjaeab/mapick/api/member/dto/request/PasswordVerifyRequest.java
@@ -1,4 +1,4 @@
-package com.gitsunjaeab.mapick.api.member.dto;
+package com.gitsunjaeab.mapick.api.member.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
