+++ MAPICK/src/main/java/com/gitsunjaeab/mapick/api/member/dto/response/MemberProfileUpdateResponse.java
@@ -1,4 +1,4 @@
-package com.gitsunjaeab.mapick.api.member.dto;
+package com.gitsunjaeab.mapick.api.member.dto.response;
 
 import com.gitsunjaeab.mapick.common.response.BaseApiResponse;
 import com.gitsunjaeab.mapick.common.response.ResponseCode;
