+++ MAPICK/src/main/java/com/gitsunjaeab/mapick/api/member/dto/response/MemberResponse.java
@@ -1,8 +1,9 @@
-package com.gitsunjaeab.mapick.api.member.dto;
+package com.gitsunjaeab.mapick.api.member.dto.response;
 
 import com.gitsunjaeab.mapick.common.response.BaseApiResponse;
 import com.gitsunjaeab.mapick.common.response.ResponseCode;
 import com.gitsunjaeab.mapick.domain.member.Member;
+import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -11,61 +12,58 @@ import java.time.OffsetDateTime;
 
 
 @Getter
-@Setter
-public class MemberProfileResponse implements BaseApiResponse {
+@AllArgsConstructor
+public class MemberResponse implements BaseApiResponse {
 
-    // 커스텀 응답 필드들
     private String code;
     private String message;
     private LocalDateTime timestamp;
-
-    // 멤버 프로필 데이터
     private MemberInfo member;
 
-    public MemberProfileResponse(String code, String message, LocalDateTime timestamp, MemberInfo member) {
-        this.code = code;
-        this.message = message;
-        this.timestamp = timestamp;
-        this.member = member;
-    }
 
     @Getter
     @Setter
     public static class MemberInfo {
         private Long id;
+        private boolean isBlacklisted;
         private String name;
         private String nickname;
         private String email;
+        private String password;
         private String loginType;
         private String provider;
         private String role;
         private String status;
         private String profileImage;
+        private OffsetDateTime lastLogin;
         private OffsetDateTime createdAt;
         private OffsetDateTime updatedAt;
         private OffsetDateTime deletedAt;
     }
 
-    public static MemberProfileResponse of(Member member) {
+    public static MemberResponse of(Member member) {
         MemberInfo memberInfo = new MemberInfo();
         memberInfo.setId(member.getId());
+        memberInfo.setBlacklisted(member.getIsBlacklisted());
         memberInfo.setName(member.getName());
         memberInfo.setNickname(member.getNickname());
         memberInfo.setEmail(member.getEmail());
+        memberInfo.setPassword(member.getPassword());
         memberInfo.setLoginType(member.getLoginType().name());
         memberInfo.setProvider(member.getProvider());
         memberInfo.setRole(member.getRole());
         memberInfo.setStatus(member.getStatus());
         memberInfo.setProfileImage(member.getProfileImage());
+        memberInfo.setLastLogin(member.getLastLogin());
         memberInfo.setCreatedAt(member.getCreatedAt());
         memberInfo.setUpdatedAt(member.getUpdatedAt());
         memberInfo.setDeletedAt(member.getDeletedAt());
 
-        return new MemberProfileResponse(
+        return new MemberResponse(
             ResponseCode.OK.getCode(),
-            "회원 프로필 조회 성공",
+            "회원 조회 성공",
             LocalDateTime.now(),
             memberInfo
         );
     }
-} 
+}
